Add view helper moving some view stuff out of the model, tidy up views, and layout
