+++ app/assets/stylesheets/application.tailwind.css
@@ -4,19 +4,19 @@
 
 
 div.field_with_errors > label {
-  @apply text-red-500;
+  @apply text-orange-700;
 }
 
 div.field_with_errors > :is(input, textarea, select) {
-  @apply border-red-500 text-red-900 placeholder-red-300;
+  @apply border-orange-700 text-orange-900 placeholder-gray-400;
 }
 
 @layer components {
   .btn-primary {
-    @apply px-4 py-2 ml-2 rounded-md bg-blue-600 text-white;
+    @apply px-4 py-2 ml-2 rounded-md bg-cyan-600 text-white;
   }
 
   .form-text-field {
-    @apply inline-block px-3 py-2 mt-2 shadow rounded-md border border-gray-200 placeholder-gray-400 outline-none;
+    @apply inline-block px-3 py-2 mt-2 shadow rounded-md  border border-gray-200 focus:border-cyan-600 focus:ring-cyan-600 placeholder-gray-400 outline-none;
   }
 }
