+++ app/views/boards/_boards.html.erb
@@ -1,3 +1,20 @@
-<table>
+<table class="p-3 border-separate border-spacing-1">
+  <tr class="px-5 bg-amber-100">
+    <th class="px-8">
+      Name
+    </th>
+    <th class="px-8">
+      Dimensions
+    </th>
+    <th class="px-8">
+      Mines
+    </th>
+    <th class="px-8">
+      Email
+    </th>
+    <th class="px-8">
+      When?
+    </th>
+  </tr>
 <%= render boards %>
 </table>
