+++ app/views/boards/index.html.erb
@@ -1,9 +1,9 @@
-<div id="boards">
-  <h1 class="font-bold text-lg mb-7">Boards</h1>
+<div id="All generated boards.">
+  <h2 class="font-bold text-lg mb-7">All generated doards</h2>
 
   <div class="px-5">
     <%= render "boards", boards: @boards %>
   </div>
 </div>
 <br>
-<%= link_to 'Back', :back, class: "btn-primary" %>
+<%= link_to 'Back', root_path, class: "btn-primary" %>
