+++ app/views/layouts/application.html.erb
@@ -12,9 +12,17 @@
   </head>
 
   <body>
+    <a href="/">
+      <header class="container mx-auto mt-8 px-5 flex">
+        <h1 class="px-2 py-2 font-bold text-lg mb7 bg-amber-200">Mysweeper</h1>
+      </header>
+    </a>
+
     <%#<main class="container mx-auto mt-28 px-5 flex"> %>
-    <main class="container mx-auto mt-28 px-5">
+    <main class="container mx-auto mt-16 px-5">
+
       <%= yield %>
     </main>
+
   </body>
 </html>
